Move to saucectl reporter
